deadcode: Optimize use of loader package
After the change to deadcode to use golang.org/x/tools/go/loader in
01514f0 the performance severly suffered. The runtime went from 0.5s
to over 2min on a codebase that I work on. I believe this was caused
by a bad use of the loader package.

I changed the code to import all packages at one time and only call
Load() once. This got the runtime of deadcode back to where it was.